+++ lib/ecto/changeset.ex
@@ -100,6 +100,18 @@ defmodule Ecto.Changeset do
   return `{:error, changeset}`, but rather raise an error at the end of the
   transaction.
 
+  ## Empty values
+  
+  Many times, the data given on cast needs to be further pruned, specially
+  regarding empty values. For example, if you are gathering data to be
+  cast from the command line or through an HTML form or any other text-based
+  format, it is likely those means cannot express nil values. For
+  those reasons, changesets include the concept of empty values, which are
+  values that will be automatically converted to the field's default value
+  on `cast/4`. Those values are stored in the changeset `empty_values` field
+  and default to `[""]`. You can also pass the `:empty_values` option to
+  `cast/4` in case you want to change how a particular `cast/4` work.
+
   ## Associations, embeds and on replace
 
   Using changesets you can work with associations as well as with embedded
@@ -233,6 +245,7 @@ defmodule Ecto.Changeset do
     * `required`     - All required fields as a list of atoms
     * `action`       - The action to be performed with the changeset
     * `types`        - Cache of the data's field types
+    * `empty_values` - A list of values to be considered empty
     * `repo`         - The repository applying the changeset (only set after a Repo function is called)
     * `repo_opts`    - A keyword list of options given to the underlying repository operation
 
@@ -260,7 +273,7 @@ defmodule Ecto.Changeset do
   defstruct valid?: false, data: nil, params: nil, changes: %{}, repo_changes: %{},
             errors: [], validations: [], required: [], prepare: [],
             constraints: [], filters: %{}, action: nil, types: nil,
-            repo: nil, repo_opts: []
+            empty_values: @empty_values, repo: nil, repo_opts: []
 
   @type t(data_type) :: %Changeset{valid?: boolean(),
                         repo: atom | nil,
@@ -459,8 +472,9 @@ defmodule Ecto.Changeset do
     raise ArgumentError, "changeset does not have types information"
   end
 
-  def cast(%Changeset{changes: changes, data: data, types: types} = changeset,
+  def cast(%Changeset{changes: changes, data: data, types: types, empty_values: empty_values} = changeset,
                       params, permitted, opts) do
+    opts = Keyword.put_new(opts, :empty_values, empty_values)
     new_changeset = cast(data, types, changes, params, permitted, opts)
     cast_merge(changeset, new_changeset)
   end