+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 v3.5 requires Elixir v1.8+.
 
+### Bug fixes
+
+  * [Ecto.Changeset] Ensure `:empty_values` in `cast/4` does not automatically propagate to following cast calls. If you want a given set of `:empty_values` to apply to all `cast/4` calls, change the value stored in `changeset.empty_values` instead
+
 ### Enhancements
 
   * [Ecto.Enum] Add `Ecto.Enum` as a custom parameterized type